requirements.txt
azure-functions==1.17.0
twilio==8.10.2
openai==1.3.2
sentry-sdk
